requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pyannote.audio==3.1.1
torch>=2.1
librosa==0.10.2.post1
soundfile==0.12.1
numpy==1.26.4
scipy==1.11.4
joblib==1.4.2
python-multipart==0.0.9
inaSpeechSegmenter==0.7.6
